added mapper and db types

+++ src/modules/blogs/router/blogsRouter.ts
@@ -5,6 +5,9 @@ import {authMiddleware} from "../../../app/config/middleware/authMiddleware";
 import {blogInputValidation} from "../validations/blogValidations";
 import {RequestBody, RequestParamsBody} from "../../common/types";
 import {BlogInputModel} from "../models/BlogInputModel";
+import {ObjectId} from "mongodb";
+import {BlogViewModel} from "../models/BlogViewModel";
+import {BlogDbType} from "../db/blog-db";
 
 export const blogsRouter = Router()
 
@@ -15,7 +18,13 @@ blogsRouter.get('/', async (req, res) => {
 })
 
 blogsRouter.get('/:blogId', async (req, res) => {
-    const foundBlogById = await blogsRepository.getBlogById(req.params.blogId)
+    const id = req.params.blogId
+
+    if (ObjectId.isValid(id)) {
+        res.sendStatus(HttpStatusCode.NOT_FOUND_404)
+        return
+    }
+    const foundBlogById = await blogsRepository.getBlogById(id)
 
     if (!foundBlogById) {
         res.sendStatus(HttpStatusCode.NOT_FOUND_404)
@@ -25,13 +34,22 @@ blogsRouter.get('/:blogId', async (req, res) => {
     res.status(HttpStatusCode.OK_200).send(foundBlogById)
 })
 
-blogsRouter.post('/', authMiddleware, blogInputValidation(), async (req: RequestBody<BlogInputModel>, res: Response) => {
-    const newBlogData: BlogInputModel = {
+blogsRouter.post('/', authMiddleware, blogInputValidation(), async (req: RequestBody<BlogInputModel>, res: Response<BlogViewModel>) => {
+    const newBlogData: BlogDbType = {
         name: req.body.name,
         description: req.body.description,
         websiteUrl: req.body.websiteUrl,
+        isMembership: false,
+        createdAt: new Date().toISOString()
     }
-    const newBlog = await blogsRepository.createNewBlog(newBlogData)
+    const newBlogId = await blogsRepository.createNewBlog(newBlogData)
+
+    const newBlog = await blogsRepository.getBlogById(newBlogId)
+
+    if (!newBlog) {
+        res.sendStatus(HttpStatusCode.NOT_FOUND_404)
+        return
+    }
 
     res.status(HttpStatusCode.CREATED_201).send(newBlog)
 })
@@ -41,6 +59,13 @@ blogsRouter.put('/:blogId', authMiddleware, blogInputValidation(), async (req: R
         name: req.body.name,
         description: req.body.description,
         websiteUrl: req.body.websiteUrl,
+    }
+
+    const foundBlog = await blogsRepository.getBlogById(req.params.blogId)
+
+    if (!foundBlog) {
+        res.sendStatus(HttpStatusCode.NOT_FOUND_404)
+        return
     }
 
     const isBlogUpdated = await blogsRepository.updateBlog(req.params.blogId, updateBlogData)
@@ -54,6 +79,13 @@ blogsRouter.put('/:blogId', authMiddleware, blogInputValidation(), async (req: R
 })
 
 blogsRouter.delete('/:blogId', authMiddleware, async (req, res) => {
+    const foundDeletedBlog = await blogsRepository.getBlogById(req.params.blogId)
+
+    if (!foundDeletedBlog) {
+        res.sendStatus(HttpStatusCode.NOT_FOUND_404)
+        return
+    }
+
     const isDeleted = await blogsRepository.deleteBlog(req.params.blogId)
 
     if (!isDeleted) {
