+++ src/modules/posts/router/postsRouter.ts
@@ -5,6 +5,9 @@ import {authMiddleware} from "../../../app/config/middleware/authMiddleware";
 import {postInputValidation} from "../validations/postValidation";
 import {RequestBody, RequestParamsBody} from "../../common/types";
 import {PostInputModel} from "../models/PostInputModel";
+import {ObjectId} from "mongodb";
+import {PostDbType} from "../db/post-db";
+import {blogsRepository} from "../../blogs/repository/blogRepository";
 
 export const postsRouter = Router()
 
@@ -15,6 +18,13 @@ postsRouter.get('/', async (req, res) => {
 })
 
 postsRouter.get('/:postId', async (req, res) => {
+    const postId = req.params.postId
+
+    if (ObjectId.isValid(postId)) {
+        res.sendStatus(HttpStatusCode.NOT_FOUND_404)
+        return
+    }
+
     const foundPost = await postsRepository.getPostById(req.params.postId)
 
     if (!foundPost) {
@@ -27,22 +37,32 @@ postsRouter.get('/:postId', async (req, res) => {
 })
 
 postsRouter.post('/', authMiddleware, postInputValidation(),  async (req: RequestBody<PostInputModel>, res: Response) => {
-    const payload: PostInputModel = {
+    const blogData = await blogsRepository.getBlogById(req.body.blogId)
+
+    if (!blogData) {
+        res.sendStatus(HttpStatusCode.NOT_FOUND_404)
+        return
+    }
+
+    const payload: PostDbType = {
         title: req.body.title,
         shortDescription: req.body.shortDescription,
         content: req.body.content,
         blogId: req.body.blogId,
+        createdAt: new Date().toISOString(),
+        blogName: blogData.name
     }
 
-    const createdPost = await postsRepository.createPost(payload)
+    const newPostId = await postsRepository.createPost(payload)
 
-    if (!createdPost) {
-        res.sendStatus(HttpStatusCode.BAD_REQUEST_400)
+    const newPost = await postsRepository.getPostById(newPostId)
 
+    if (!newPost) {
+        res.sendStatus(HttpStatusCode.NOT_FOUND_404)
         return
     }
 
-    res.status(HttpStatusCode.CREATED_201).send(createdPost)
+    res.status(HttpStatusCode.CREATED_201).send(newPost)
 })
 
 postsRouter.put('/:postId', authMiddleware, postInputValidation(),  async (req: RequestParamsBody<{ postId: string }, PostInputModel>, res: Response) => {
@@ -52,6 +72,14 @@ postsRouter.put('/:postId', authMiddleware, postInputValidation(),  async (req: 
         content: req.body.content,
         blogId: req.body.blogId,
     }
+
+    const foundPost = await postsRepository.getPostById(req.params.postId)
+
+    if (!foundPost) {
+        res.sendStatus(HttpStatusCode.NOT_FOUND_404)
+        return
+    }
+
 
     const isPostUpdated = await postsRepository.updatePost(payload, req.params.postId)
 
@@ -65,6 +93,14 @@ postsRouter.put('/:postId', authMiddleware, postInputValidation(),  async (req: 
 })
 
 postsRouter.delete('/:postId', authMiddleware, async (req, res) => {
+    const foundDeletedPost = await postsRepository.getPostById(req.params.blogId)
+
+    if (!foundDeletedPost) {
+        res.sendStatus(HttpStatusCode.NOT_FOUND_404)
+        return
+    }
+
+
     const isDeleted = await postsRepository.deletePostById(req.params.postId)
 
     if (!isDeleted) {
